agent-cluster/web_app_async.py
#!/usr/bin/env python3
"""
Agent Cluster 异步 Web 界面 (FastAPI 版本)
性能优化版本，保留原有同步版本
"""

from fastapi import FastAPI, HTTPException, Header, Depends
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, Dict, Any
import subprocess
from datetime import datetime
from pathlib import Path

# 导入配置和认证
from utils.config_loader import config
from utils.auth import jwt_auth, require_auth

app = FastAPI(
    title="AI 产品开发智能体",
    description="完整 6 阶段开发流程 | 10 个专业 Agent | 质量门禁 | 钉钉通知",
    version="2.2.0"
)

# CORS 配置
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# 请求/响应模型
class LoginRequest(BaseModel):
    username: str
    password: str

class TaskSubmit(BaseModel):
    requirement: str
    project: str = "default"

# 健康检查
@app.get("/api/health")
async def health_check():
    """健康检查端点"""
    return {
        'status': 'healthy',
        'timestamp': datetime.now().isoformat(),
        'service': 'agent-cluster',
        'version': '2.2.0'
    }

# 状态检查
@app.get("/api/status")
async def get_status():
    """获取系统状态"""
    try:
        orch = bool(subprocess.run(['pgrep', '-f', 'orchestrator.py'], 
                   capture_output=True, text=True, timeout=5).stdout.strip())
        return {
            "status": "running" if orch else "stopped",
            "timestamp": datetime.now().isoformat(),
            "active_workflows": 0,
            "completed_today": 0,
            "failed_today": 0,
            "agents_ready": 10,
            "dingtalk_enabled": bool(config.dingtalk_webhook)
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# 登录
@app.post("/api/login")
async def login(request: LoginRequest):
    """用户登录"""
    # 简单验证（实际应从数据库验证）
    if request.username == "admin" and request.password == "admin123":
        token = jwt_auth.create_token(request.username)
        return {"success": True, "token": token}
    raise HTTPException(status_code=401, detail="用户名或密码错误")

# 获取工作流
@app.get("/api/workflows")
async def get_workflows(user: dict = Depends(require_auth)):
    """获取工作流列表"""
    # TODO: 实现工作流获取
    return {"workflows": []}

# 提交任务
@app.post("/api/submit")
async def submit_task(task: TaskSubmit, user: dict = Depends(require_auth)):
    """提交新任务"""
    if not task.requirement:
        raise HTTPException(status_code=400, detail="需求不能为空")
    
    # TODO: 实现任务提交
    return {"success": True, "message": "任务已提交"}

# 获取 Agent 列表
@app.get("/api/agents")
async def get_agents():
    """获取 Agent 列表"""
    return {
        "executors": [
            {"name": "Product Manager", "role": "产品经理", "phase": "Phase 1"},
            {"name": "Tech Lead", "role": "技术负责人", "phase": "Phase 2"},
            {"name": "Designer", "role": "设计师", "phase": "Phase 2"},
            {"name": "DevOps", "role": "运维工程师", "phase": "Phase 2/6"},
            {"name": "Codex", "role": "后端专家", "phase": "Phase 3"},
            {"name": "Claude-Code", "role": "前端专家", "phase": "Phase 3"},
            {"name": "Tester", "role": "测试工程师", "phase": "Phase 4"}
        ],
        "reviewers": [
            {"name": "Codex Reviewer", "role": "逻辑审查", "phase": "Phase 5"},
            {"name": "Gemini Reviewer", "role": "安全审查", "phase": "Phase 5"},
            {"name": "Claude Reviewer", "role": "基础审查", "phase": "Phase 5"}
        ]
    }

# 启动命令
if __name__ == '__main__':
    import uvicorn
    uvicorn.run(
        app,
        host=config.host,
        port=config.port,
        reload=config.debug
    )
