agent-cluster/requirements.txt
# Agent Cluster 依赖

# 核心依赖
requests>=2.28.0
PyJWT>=2.6.0
psutil>=5.9.0

# 异步框架（可选）
fastapi>=0.100.0
uvicorn>=0.23.0

# 缓存（可选）
redis>=4.5.0

# 数据库（可选）
sqlalchemy>=2.0.0

# 开发依赖
pytest>=7.0.0
black>=23.0.0
